创建了几个基础接口feat(service): 添加维护记录更新和用户删除功能
- 在 DataETableServiceImpl 中添加维护记录保存逻辑
- 在 IMaintainTableService 中添加更新维护记录接口
- 在 MaintainTableController 中添加更新维护记录接口
- 在 MaintainTableServiceImpl 中实现更新维护记录方法
- 在 IUsersService 中添加删除用户接口
- 在 UsersController 中添加删除用户接口
- 在 UsersServiceImpl 中实现删除用户方法
- 新增 MaintainTableDTO 类用于维护记录更新

+++ src/main/java/com/example/V1/entity/DataETable.java
@@ -64,7 +64,7 @@ public class DataETable implements Serializable {
      */
     @JsonProperty("eData")
     @TableField("e_data")
-    private int eData;
+    private double eData;
 
 
 }
